Move Arr tests to the right location
