+++ tests/Support/SupportHelpersTest.php
@@ -6,7 +6,6 @@ use stdClass;
 use ArrayAccess;
 use Mockery as m;
 use RuntimeException;
-use Illuminate\Support\Arr;
 use PHPUnit\Framework\TestCase;
 use Illuminate\Support\Optional;
 use Illuminate\Contracts\Support\Htmlable;
@@ -16,196 +15,6 @@ class SupportHelpersTest extends TestCase
     protected function tearDown(): void
     {
         m::close();
-    }
-
-    public function testArrayDot()
-    {
-        $array = Arr::dot(['name' => 'taylor', 'languages' => ['php' => true]]);
-        $this->assertEquals($array, ['name' => 'taylor', 'languages.php' => true]);
-    }
-
-    public function testArrayGet()
-    {
-        $array = ['names' => ['developer' => 'taylor']];
-        $this->assertEquals('taylor', Arr::get($array, 'names.developer'));
-        $this->assertEquals('dayle', Arr::get($array, 'names.otherDeveloper', 'dayle'));
-        $this->assertEquals('dayle', Arr::get($array, 'names.otherDeveloper', function () {
-            return 'dayle';
-        }));
-    }
-
-    public function testArrayHas()
-    {
-        $array = ['names' => ['developer' => 'taylor']];
-        $this->assertTrue(Arr::has($array, 'names'));
-        $this->assertTrue(Arr::has($array, 'names.developer'));
-        $this->assertFalse(Arr::has($array, 'foo'));
-        $this->assertFalse(Arr::has($array, 'foo.bar'));
-    }
-
-    public function testArraySet()
-    {
-        $array = [];
-        Arr::set($array, 'names.developer', 'taylor');
-        $this->assertEquals('taylor', $array['names']['developer']);
-    }
-
-    public function testArrayForget()
-    {
-        $array = ['names' => ['developer' => 'taylor', 'otherDeveloper' => 'dayle']];
-        Arr::forget($array, 'names.developer');
-        $this->assertFalse(isset($array['names']['developer']));
-        $this->assertTrue(isset($array['names']['otherDeveloper']));
-
-        $array = ['names' => ['developer' => 'taylor', 'otherDeveloper' => 'dayle', 'thirdDeveloper' => 'Lucas']];
-        Arr::forget($array, ['names.developer', 'names.otherDeveloper']);
-        $this->assertFalse(isset($array['names']['developer']));
-        $this->assertFalse(isset($array['names']['otherDeveloper']));
-        $this->assertTrue(isset($array['names']['thirdDeveloper']));
-
-        $array = ['names' => ['developer' => 'taylor', 'otherDeveloper' => 'dayle'], 'otherNames' => ['developer' => 'Lucas', 'otherDeveloper' => 'Graham']];
-        Arr::forget($array, ['names.developer', 'otherNames.otherDeveloper']);
-        $expected = ['names' => ['otherDeveloper' => 'dayle'], 'otherNames' => ['developer' => 'Lucas']];
-        $this->assertEquals($expected, $array);
-    }
-
-    public function testArrayPluckWithArrayAndObjectValues()
-    {
-        $array = [(object) ['name' => 'taylor', 'email' => 'foo'], ['name' => 'dayle', 'email' => 'bar']];
-        $this->assertEquals(['taylor', 'dayle'], Arr::pluck($array, 'name'));
-        $this->assertEquals(['taylor' => 'foo', 'dayle' => 'bar'], Arr::pluck($array, 'email', 'name'));
-    }
-
-    public function testArrayPluckWithNestedKeys()
-    {
-        $array = [['user' => ['taylor', 'otwell']], ['user' => ['dayle', 'rees']]];
-        $this->assertEquals(['taylor', 'dayle'], Arr::pluck($array, 'user.0'));
-        $this->assertEquals(['taylor', 'dayle'], Arr::pluck($array, ['user', 0]));
-        $this->assertEquals(['taylor' => 'otwell', 'dayle' => 'rees'], Arr::pluck($array, 'user.1', 'user.0'));
-        $this->assertEquals(['taylor' => 'otwell', 'dayle' => 'rees'], Arr::pluck($array, ['user', 1], ['user', 0]));
-    }
-
-    public function testArrayPluckWithNestedArrays()
-    {
-        $array = [
-            [
-                'account' => 'a',
-                'users' => [
-                    ['first' => 'taylor', 'last' => 'otwell', 'email' => 'taylorotwell@gmail.com'],
-                ],
-            ],
-            [
-                'account' => 'b',
-                'users' => [
-                    ['first' => 'abigail', 'last' => 'otwell'],
-                    ['first' => 'dayle', 'last' => 'rees'],
-                ],
-            ],
-        ];
-
-        $this->assertEquals([['taylor'], ['abigail', 'dayle']], Arr::pluck($array, 'users.*.first'));
-        $this->assertEquals(['a' => ['taylor'], 'b' => ['abigail', 'dayle']], Arr::pluck($array, 'users.*.first', 'account'));
-        $this->assertEquals([['taylorotwell@gmail.com'], [null, null]], Arr::pluck($array, 'users.*.email'));
-    }
-
-    public function testArrayExcept()
-    {
-        $array = ['name' => 'taylor', 'age' => 26];
-        $this->assertEquals(['age' => 26], Arr::except($array, ['name']));
-        $this->assertEquals(['age' => 26], Arr::except($array, 'name'));
-
-        $array = ['name' => 'taylor', 'framework' => ['language' => 'PHP', 'name' => 'Laravel']];
-        $this->assertEquals(['name' => 'taylor'], Arr::except($array, 'framework'));
-        $this->assertEquals(['name' => 'taylor', 'framework' => ['name' => 'Laravel']], Arr::except($array, 'framework.language'));
-        $this->assertEquals(['framework' => ['language' => 'PHP']], Arr::except($array, ['name', 'framework.name']));
-    }
-
-    public function testArrayOnly()
-    {
-        $array = ['name' => 'taylor', 'age' => 26];
-        $this->assertEquals(['name' => 'taylor'], Arr::only($array, ['name']));
-        $this->assertEmpty(Arr::only($array, ['nonExistingKey']));
-    }
-
-    public function testArrayCollapse()
-    {
-        $array = [[1], [2], [3], ['foo', 'bar'], collect(['baz', 'boom'])];
-        $this->assertEquals([1, 2, 3, 'foo', 'bar', 'baz', 'boom'], Arr::collapse($array));
-    }
-
-    public function testArrayDivide()
-    {
-        $array = ['name' => 'taylor'];
-        [$keys, $values] = Arr::divide($array);
-        $this->assertEquals(['name'], $keys);
-        $this->assertEquals(['taylor'], $values);
-    }
-
-    public function testArrayFirst()
-    {
-        $array = ['name' => 'taylor', 'otherDeveloper' => 'dayle'];
-        $this->assertEquals('dayle', Arr::first($array, function ($value) {
-            return $value == 'dayle';
-        }));
-    }
-
-    public function testArrayLast()
-    {
-        $array = [100, 250, 290, 320, 500, 560, 670];
-        $this->assertEquals(670, Arr::last($array, function ($value) {
-            return $value > 320;
-        }));
-    }
-
-    public function testArrayPluck()
-    {
-        $data = [
-            'post-1' => [
-                'comments' => [
-                    'tags' => [
-                        '#foo', '#bar',
-                    ],
-                ],
-            ],
-            'post-2' => [
-                'comments' => [
-                    'tags' => [
-                        '#baz',
-                    ],
-                ],
-            ],
-        ];
-
-        $this->assertEquals([
-            0 => [
-                'tags' => [
-                    '#foo', '#bar',
-                ],
-            ],
-            1 => [
-                'tags' => [
-                    '#baz',
-                ],
-            ],
-        ], Arr::pluck($data, 'comments'));
-
-        $this->assertEquals([['#foo', '#bar'], ['#baz']], Arr::pluck($data, 'comments.tags'));
-        $this->assertEquals([null, null], Arr::pluck($data, 'foo'));
-        $this->assertEquals([null, null], Arr::pluck($data, 'foo.bar'));
-    }
-
-    public function testArrayPrepend()
-    {
-        $array = Arr::prepend(['one', 'two', 'three', 'four'], 'zero');
-        $this->assertEquals(['zero', 'one', 'two', 'three', 'four'], $array);
-
-        $array = Arr::prepend(['one' => 1, 'two' => 2], 0, 'zero');
-        $this->assertEquals(['zero' => 0, 'one' => 1, 'two' => 2], $array);
-    }
-
-    public function testArrayFlatten()
-    {
-        $this->assertEquals(['#foo', '#bar', '#baz'], Arr::flatten([['#foo', '#bar'], ['#baz']]));
     }
 
     public function testE()
@@ -499,76 +308,6 @@ class SupportHelpersTest extends TestCase
         ], $data);
     }
 
-    public function testArraySort()
-    {
-        $array = [
-            ['name' => 'baz'],
-            ['name' => 'foo'],
-            ['name' => 'bar'],
-        ];
-
-        $this->assertEquals([
-            ['name' => 'bar'],
-            ['name' => 'baz'],
-            ['name' => 'foo'], ],
-        array_values(Arr::sort($array, function ($v) {
-            return $v['name'];
-        })));
-    }
-
-    public function testArraySortRecursive()
-    {
-        $array = [
-            [
-                'foo',
-                'bar',
-                'baz',
-            ],
-            [
-                'baz',
-                'foo',
-                'bar',
-            ],
-        ];
-
-        $assumedArray = [
-            [
-                'bar',
-                'baz',
-                'foo',
-            ],
-            [
-                'bar',
-                'baz',
-                'foo',
-            ],
-        ];
-
-        $this->assertEquals($assumedArray, Arr::sortRecursive($array));
-    }
-
-    public function testArrayWhere()
-    {
-        $array = ['a' => 1, 'b' => 2, 'c' => 3, 'd' => 4, 'e' => 5, 'f' => 6, 'g' => 7, 'h' => 8];
-        $this->assertEquals(['b' => 2, 'd' => 4, 'f' => 6, 'h' => 8], Arr::where(
-            $array,
-            function ($value, $key) {
-                return $value % 2 === 0;
-            }
-        ));
-    }
-
-    public function testArrayWrap()
-    {
-        $string = 'a';
-        $array = ['a'];
-        $object = new stdClass;
-        $object->value = 'a';
-        $this->assertEquals(['a'], Arr::wrap($string));
-        $this->assertEquals($array, Arr::wrap($array));
-        $this->assertEquals([$object], Arr::wrap($object));
-    }
-
     public function testHead()
     {
         $array = ['a', 'b', 'c'];
@@ -607,19 +346,6 @@ class SupportHelpersTest extends TestCase
             SupportTestTraitThree::class => SupportTestTraitThree::class,
         ],
         class_uses_recursive(SupportTestClassThree::class));
-    }
-
-    public function testArrayAdd()
-    {
-        $this->assertEquals(['surname' => 'Mövsümov'], Arr::add([], 'surname', 'Mövsümov'));
-        $this->assertEquals(['developer' => ['name' => 'Ferid']], Arr::add([], 'developer.name', 'Ferid'));
-    }
-
-    public function testArrayPull()
-    {
-        $developer = ['firstname' => 'Ferid', 'surname' => 'Mövsümov'];
-        $this->assertEquals('Mövsümov', Arr::pull($developer, 'surname'));
-        $this->assertEquals(['firstname' => 'Ferid'], $developer);
     }
 
     public function testTap()
